pyproject.toml
[tool.poetry]
name = "discord-bot-the-eternal-gem"
version = "0.1.0"
description = ""
authors = ["lesteenman <lesteenman@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^1.6.0"
loguru = "^0.5.3"
PyYAML = "^5.4.1"

[tool.poetry.dev-dependencies]
pytest = "^6.2.2"
pytest-mock = "^3.5.1"
flake8 = "^3.9.0"
pytest-asyncio = "^0.14.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
